CMakelists: Fix FTBS on Debian sid (+ likely other gtk3 platforms)
On platforms using gtk3, invoke  FIND_PACKAGE(wxWidgets) with proper
options to handle those who have mgtk2 as default, like Debian sid.
Also add the __WXGTK3__ symbol which seems to be missing from the
wx-config output on Debian.

Here is plenty of room for clean-ups; the condition WIN32 or
APPLE or QT_ANDROID comes to mind.