CLI implementing plugin linkage check + install, uninstall, list
Adds the program opencpn-cli which can be used to check that a
plugin can be loaded without missing runtime dependencies.

On windows, the installer places the program in
.../OpenCPN/bin/opencpnc-cli. In order to be used, it must be
moved to .../OpenCPN i. e., the same directory which holds the
regular opencpn GUI app. It's not that useful on Windows anyway
since plugin dependencies there are checked in build time.

The program depends on a shim of the plugin API which only provides
linkage. It has no functionality and thus no dependencies.

The install, uninstall and list basically makes part of the
PluginHandler interface available on the command line.

cli:  Implement load, install, uninstall, list.